`uint16_t tileIdx` type matches `options.baseTileIDs`
Tile IDs wrap at 8 bits, but that's handled by `uint8_t tileID`

+++ src/gfx/process.cpp
@@ -691,7 +691,7 @@ static void outputUnoptimizedMaps(
 	autoOpenPath(options.attrmap, attrmapOutput);
 	autoOpenPath(options.palmap, palmapOutput);
 
-	uint8_t tileIdx = 0;
+	uint16_t tileIdx = 0;
 	uint8_t bank = 0;
 	for (AttrmapEntry const &attr : attrmap) {
 		// The update-increment logic at the end of this loop may increment `bank` from 1 to 2,
